Create utils file for helper functions and variables
