+++ frontend/src/pages/UpdateEvent/index.tsx
@@ -1,11 +1,12 @@
-import {ChangeEvent, FormEvent, useState, useEffect} from 'react';
+import { FormEvent, useEffect } from 'react';
 import axios from 'axios';
 import EventForm from '../../components/EventForm';
+import { useParams } from 'react-router-dom';
 import './index.css';
-import {useParams} from "react-router-dom";
+import {useForm} from "../../resources/formUtils.ts";
 
 export default function UpdateEvent(): JSX.Element {
-    const {eventId} = useParams<{ eventId: string }>();
+    const { eventId } = useParams<{ eventId: string }>();
 
     const initialFormData = {
         name: '',
@@ -17,28 +18,19 @@ export default function UpdateEvent(): JSX.Element {
         usersWhoDownvoted: [],
     };
 
-    const [formData, setFormData] = useState(initialFormData);
+    const { formData, handleChange, isValid, setFormData } = useForm(initialFormData);
 
     useEffect(() => {
         axios
             .get(`/api/events/${eventId}`)
             .then((response): void => {
                 const eventToUpdate = response.data;
-                setFormData(eventToUpdate);
+                setFormData(eventToUpdate); // Set the form data
             })
             .catch((error): void => {
                 console.error('Error fetching event details:', error);
             });
     }, [eventId]);
-
-    const handleChange = (e: ChangeEvent<HTMLInputElement>): void => {
-        const {name, value} = e.target;
-        setFormData((prevData) => ({...prevData, [name]: value}));
-    };
-
-    const isValid: boolean = Object.values(formData).every(
-        (value) => typeof value === 'string' && value.trim() !== '' && value.length >= 3
-    );
 
     const handleSubmit = (e: FormEvent): void => {
         e.preventDefault();
@@ -56,14 +48,12 @@ export default function UpdateEvent(): JSX.Element {
     const handleDelete = async (): Promise<void> => {
         try {
             await axios.delete(`/api/events/${eventId}`);
-
         } catch (error) {
             console.error('Error during delete:', error);
         }
     };
 
     return (
-                <EventForm formData={formData} isValid={isValid}
-                           handleDelete={handleDelete} handleSubmit={handleSubmit} onChange={handleChange}/>
+        <EventForm formData={formData} isValid={isValid} handleDelete={handleDelete} handleSubmit={handleSubmit} onChange={handleChange} />
     );
 }
